backend/requirements.txt
numpy
pyserial==3.5
opencv-python
matplotlib
flask
flask-cors